Split `owner` and `team` into separate models
