+++ src/github.rs
@@ -93,3 +93,12 @@ pub fn token(token: String) -> Token {
         token_type: String::new(),
     }
 }
+
+pub fn team_url(login: &str) -> String {
+    let mut login_pieces = login.split(':');
+    login_pieces.next();
+    format!(
+        "https://github.com/{}",
+        login_pieces.next().expect("org failed"),
+    )
+}
