+++ src/models/mod.rs
@@ -6,8 +6,9 @@ pub use download::VersionDownload;
 pub use self::follow::Follow;
 pub use self::keyword::{CrateKeyword, Keyword};
 pub use self::krate::{Crate, CrateDownload, NewCrate};
-pub use owner::{CrateOwner, NewTeam, Owner, OwnerKind, Team};
+pub use self::owner::{CrateOwner, Owner, OwnerKind};
 pub use self::rights::Rights;
+pub use self::team::{NewTeam, Team};
 pub use user::{Email, NewUser, User};
 pub use token::ApiToken;
 pub use version::{NewVersion, Version};
@@ -19,4 +20,6 @@ mod category;
 mod follow;
 mod keyword;
 pub mod krate;
+mod owner;
 mod rights;
+mod team;
